+++ src/dipdup/config.py
@@ -952,7 +952,6 @@ default_hooks = {
 @dataclass
 class AdvancedConfig:
     reindex: Dict[ReindexingReasonC, ReindexingAction] = field(default_factory=dict)
-    oneshot: bool = False
     postpone_jobs: bool = False
     early_realtime: bool = False
     merge_subscriptions: bool = False
@@ -1013,12 +1012,9 @@ class DipDupConfig:
     def oneshot(self) -> bool:
         syncable_indexes = tuple(c for c in self.indexes.values() if not isinstance(c, HeadIndexConfig))
         oneshot_indexes = tuple(c for c in syncable_indexes if c.last_level)
-        if not oneshot_indexes:
-            return False
-        elif len(oneshot_indexes) == len(syncable_indexes):
+        if len(oneshot_indexes) == len(syncable_indexes):
             return True
-        else:
-            raise ConfigurationError('Either all or none of indexes can have `last_level` field set')
+        return False
 
     @classmethod
     def load(